Funcionalidades Admin en product con nuevos estilos. Scrollbar nueva

+++ frontend/src/components/ProductItem.js
@@ -14,68 +14,70 @@ import 'react-toastify/dist/ReactToastify.css';
 
 const ProductItem = ({ data, functionReload, handleClose }) => {
 
-  let { _id, name, price, image, stock} = data;
+  let { _id, name, price, image, stock } = data;
   const dispatch = useDispatch();
   const [opcion, setOpcion] = useState(10);
-  const [inputname, setInputname] = useState(name);
-  const [inputstock, setInputstock] = useState(stock);
-  const [inputprice, setInputprice] = useState(price);
-  
+  const [inputName, setInputname] = useState(name);
+  const [inputStock, setInputstock] = useState(stock);
+  const [inputPrice, setInputprice] = useState(price);
+  const [inputImage, setInputImage] = useState(image);
+
   const user = useSelector(store => store.userReducer.user);
-  
+
   const handleChange = (event) => {
     setOpcion(event.target.value);
   };
 
-  async function deleteProduct(event) {
-    await dispatch(productActions.deleteProduct(_id))
-      .then(response => {
-        if (response.data.success) {
-          toast.success(response.data.message)
-        } else {
-          toast.error(response.data.message)
-        }
-      })
-    functionReload()
+  async function deleteProduct() {
+    if (window.confirm(`¿Seguro que desea eliminar ${name}?`)) {
+      await dispatch(productActions.deleteProduct(_id))
+        .then(response => {
+          if (response.data.success) {
+            toast.success(response.data.message);
+          } else {
+            toast.error(response.data.message);
+          }
+        })
+      functionReload();
+    }
   };
 
   async function modifyProduct(event) {
     const modifiedProduct = {
-      name: inputname,
-      price: inputprice,
-      stock: inputstock,
+      name: inputName,
+      price: inputPrice,
+      stock: inputStock,
+      image: inputImage,
       id: _id
     }
-    await dispatch(productActions.modifyProduct(modifiedProduct))
-    functionReload()
-    handleClose()
+    await dispatch(productActions.modifyProduct(modifiedProduct));
+    functionReload();
   };
 
   return (
     <div className="box-modal w-4/5 sm:w-96">
-      <div className='more-border'>
+      <div className='more-border scroll-products'>
         <img className='image-modal w-24 sm:w-32' src={image} alt="product" />
-        {user?.user.role === 'admin' ?
-          <input type='textarea' defaultValue={name} value={inputname} onChange={(event) => setInputname(event.target.value)} className='more-name text-sm font-products' />
-          :
-          <h4 className="text-lg font-products">{name}</h4>
-        }
-
         <div className='flex flex-col lg:flex-row justify-around items-center w-full'>
           {user?.user.role === 'admin' ?
-            <>
-              <p>Stock:</p>
-              <input type='textarea' defaultValue={stock} onChange={(event) => setInputstock(event.target.value)} className='more-stock' />
-              <p>Precio: </p>
-              <input type='textarea' defaultValue={price} onChange={(event) => setInputprice(event.target.value)} className='more-price'></input>
-            </>
+            <div className='more-data my-2 flex flex-col items-center justify-around'>
+              <label for="product-name-input" className='text-lg'>Nombre:</label>
+              <input type='text' defaultValue={name} value={inputName} onChange={(event) => setInputname(event.target.value)} className='w-3/4 text-lg font-products my-1' name='product-name-input' id='product-name-input' />
+              <label for="product-stock-input" className='mt-2 text-lg'>Stock:</label>
+              <input className='my-1 rounded p-2 w-3/4 my-1' type='number' defaultValue={stock} onChange={(event) => setInputstock(event.target.value)} name='product-stock-input' id='product-stock-input' />
+              <label for="product-price-input" className='mt-2 text-lg'>Precio: </label>
+              <input className='my-1 rounded p-2 w-3/4 my-1' type='number' defaultValue={price} onChange={(event) => setInputprice(event.target.value)} name='product-price-input' id='product-price-input' />
+              <label for="product-img-input" className='mt-2 text-lg'>Imagen (URL): </label>
+              <input className='my-1 rounded p-2 w-3/4 my-1' type='text' defaultValue={image} onChange={(event) => setInputImage(event.target.value)} name='product-img-input' id='product-img-input' />
+            </div>
             :
             <>
+              <h4 className="text-lg font-products">{name}</h4>
               <div className='more-data my-2'>
                 <p>Disponible:</p>
                 <p>{stock} unidades</p>
               </div>
-              {data.variations.length>0 &&
+              {data.variations.length > 0 &&
                 <FormControl color='warning' className='more-select' sx={{ m: 2, minWidth: 120 }} size="small">
                   <InputLabel color='warning' id="demo-select-small">Variedades:</InputLabel>
                   <Select
@@ -88,7 +90,7 @@ const ProductItem = ({ data, functionReload, handleClose }) => {
                     onChange={handleChange}
                   >
                     {data.variations.map((item, i) => (
-                      <MenuItem color='warning' key={i} value={(i+1)*10}>{item}</MenuItem>
+                      <MenuItem color='warning' key={i} value={(i + 1) * 10}>{item}</MenuItem>
                     ))}
                   </Select>
                 </FormControl>
@@ -101,10 +103,9 @@ const ProductItem = ({ data, functionReload, handleClose }) => {
         {user?.user.role === 'admin' ?
           <div className='admin-container'>
             <h3 className='admin-title'>« Administrador »</h3>
-            <h3 className='admin-text'>¿Que cambios desea realizar en el producto?</h3>
-            <div >
-              <button className="  button-admin" onClick={modifyProduct}> Modificarlo <ModeEditIcon /></button>
-              <button className=" button-admin" onClick={deleteProduct}>Eliminarlo<DeleteIcon /></button>
+            <div className='w-full flex flex-col sm:flex-row justify-around items-center' >
+              <button className="  button-admin" onClick={modifyProduct}>Guardar<ModeEditIcon /></button>
+              <button className=" button-admin" onClick={deleteProduct}>Eliminar<DeleteIcon /></button>
             </div>
           </div>
           : user?.user.role === 'user' ?
